==> community/sunny_metrics/02_label_app_contracts.ipynb <==
# 02 Label apps based on the type of contract

This notebook filters the apps based on what type of contract they have:

- token_contract: ERC20 token contracts. Metrics are derived from token transfer event data (Dune).

- dapp_contract: Regular smart contracts. Metrics are derived from raw transaction data (Goldsky). 

- factory_contract: Contracts used to generate other contracts. Metrics pendings.

- trace_contract: Something else, but that has events in traces.

from google.cloud import bigquery
import os
import pandas as pd

os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = "../../oso_gcp_credentials.json"
client = bigquery.Client()
PROJECT = 'opensource-observer'
CURRENT_CHAINS = ['Optimism', 'Base', 'Zora', 'Mode', 'Frax', 'Metal']

all_apps = pd.read_csv('data/apps/applications_reviewed.csv')

all_apps = all_apps[all_apps['project_type'] != 'Creator']
all_apps['chain'].fillna('', inplace=True)
all_apps['address'].fillna('', inplace=True)
all_apps = all_apps[~all_apps['chain'].isin(['All Superchain', ''])]
all_apps = all_apps[all_apps['address'].apply(len)==42]
all_apps.set_index('uuid', inplace=True)

all_apps.tail(1)

INVALID_CONTRACTS = {
    "0x4e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec67": "safe",
    "0x4200000000000000000000000000000000000020": "eas",
    "0x4200000000000000000000000000000000000021": "eas",
}

CONTRACT_LABELS = {
    uuid: {
        'address': row['address'],
        'chain': row['chain'],
        'label': 'unknown' if row['address'] not in INVALID_CONTRACTS else 'invalid_contract'
    }
    for uuid, row in all_apps.iterrows()
}

## Part 1. Token contracts

token_list = pd.read_csv('data/raw_metric_data/dune_token_list.csv', low_memory=False)

tokens = {
    'Base': [],
    'Optimism': [],    
    'Zora': []
}

for _,row in token_list.iterrows():
    chain = row['blockchain'].title()
    addresses = row['token_addresses']
    addresses = addresses[1:-1]
    for a in addresses.split(' '):
        if len(a) == 42:
            tokens[chain].append(a.lower())

for uuid,app in CONTRACT_LABELS.items():
    
    if app['label'] != 'unknown':
        continue
    
    chain = app['chain']
    address = app['address']
    
    if chain not in CURRENT_CHAINS:
        CONTRACT_LABELS[uuid]['label'] = f"pending_{chain.lower()}"
        print(CONTRACT_LABELS[uuid])
        continue
        
    if tokens.get(chain):
        if address in tokens[chain]:
            CONTRACT_LABELS[uuid]['label'] = 'token_contract'
            print(CONTRACT_LABELS[uuid])
            continue

## Part 2. Determine best way to handle other contracts

def oso_lookup_query(uuid_contract_chain_tuples):
    
    list_of_contracts = [x[1] for x in uuid_contract_chain_tuples if isinstance(x[1], str)]
    list_of_contracts_str = "'" + "','".join(list_of_contracts) + "'"
    
    query = f"""
        with contracts as (
            select 
                to_address as address,
                upper(chain) as chain,
                txns as num_events,
                'transactions' as data_source
            from `{PROJECT}.static_data_sources.sunny_contract_lookup`
            union all
            select 
                address,
                upper(chain) as chain,
                count_transactions as num_events,
                'traces' as data_source
            from `{PROJECT}.static_data_sources.sunny_trace_lookup`
            union all
            select
                factory_address as address,
                upper(chain) as chain,
                count_contracts as num_events,
                'factories' as data_source
            from `{PROJECT}.static_data_sources.sunny_factory_lookup`
        )
        select
            address,
            chain,
            sum(case when data_source = 'transactions' then num_events else 0 end) as count_transactions,
            sum(case when data_source = 'traces' then num_events else 0 end) as count_traces,
            sum(case when data_source = 'factories' then num_events else 0 end) as count_factory_deploys
        from contracts
        where address in ({list_of_contracts_str})
        group by address, chain
    """
    
    result = client.query(query)
    dataframe = result.to_dataframe()

    labeled_contracts = []
    for (uuid, address, chain) in uuid_contract_chain_tuples:
        
        item = {'uuid': uuid, 'address': address, 'chain': chain}
    
        if not isinstance(chain, str) or not isinstance(address, str):
            continue

        temp_ = dataframe[(dataframe['address']==address) & (dataframe['chain']==chain.upper())]
        
        if not len(temp_):
            labeled_contracts.append({**item, 'label': 'unknown'})
            continue

        c = temp_.set_index('address').iloc[0]

        # look for factories
        if c.count_factory_deploys > c.count_traces:
            labeled_contracts.append({**item, 'label': 'factory_contract'})
            continue

        # low activity        
        if c.count_transactions < 10 and c.count_traces < 10:
            if c.count_factory_deploys:
                labeled_contracts.append({**item, 'label': 'factory_contract'})
            else:            
                if c.count_factory_deploys:
                    labeled_contracts.append({**item, 'label': 'inactive_contract'})
            continue

        # only traces cases
        if not c.count_transactions:
            labeled_contracts.append({**item, 'label': 'trace_contract'})
            continue

        # only transactions cases
        if not c.count_traces:
            labeled_contracts.append({**item, 'label': 'dapp_contract'})
            continue

        # decided based on ratio of transactions to traces
        if c.count_traces / c.count_transactions >= 1.5:
            labeled_contracts.append({**item, 'label': 'trace_contract'})
        else:
            labeled_contracts.append({**item, 'label': 'dapp_contract'})  
            
    return labeled_contracts

lookup_results = oso_lookup_query([
    (uuid, app['address'], app['chain'])
    for uuid, app in CONTRACT_LABELS.items()
    if app['label'] == 'unknown'
])

for res in lookup_results:
    uuid = res['uuid']
    label = res['label']
    CONTRACT_LABELS[uuid]['label'] = label

## Part 3. Check what's left and dump it

df_labels = pd.DataFrame(CONTRACT_LABELS).T
df_labels.rename(columns={'label': 'contract_type'}, inplace=True)
df_labels.tail(1)

cols = ['recipient', 'project_type', 'category']
all_apps[cols].join(df_labels).to_csv('data/apps/project_apps_labeled.csv')
