==> community/sunny_metrics/05_derive_creator_metrics.ipynb <==
# Creator Metrics

This script processes a CSV from [this Dune Query](https://dune.com/queries/4087193) and derives metrics the following NFT creator metrics:

- num drops
- num unique minters
- num txns
- USD value of txns

from dotenv import load_dotenv
from dune_client.types import QueryParameter
from dune_client.client import DuneClient
from dune_client.query import QueryBase
from google.cloud import bigquery
import json
import os
import pandas as pd

load_dotenv()

PROJECT = 'opensource-observer'
os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = "../../oso_gcp_credentials.json"
client = bigquery.Client()


DUNE_API_KEY = os.getenv('DUNE_API_KEY')
dune = DuneClient(DUNE_API_KEY)

## Part 1. Parse the creator data

apps = pd.read_csv('data/apps/creator_apps.csv', index_col=0)
apps = apps.drop(columns=['chain']).drop_duplicates()
uuid_to_address_mapper = apps.set_index('uuid')['address'].to_dict()
uuid_list = list(apps['uuid'].unique())
address_list = list(apps['address'].unique())
apps.tail(1)

all_apps = pd.read_csv('data/apps/applications_reviewed.csv')
all_apps = all_apps[all_apps['uuid'].isin(uuid_list)]
uuid_to_profile_mapper = all_apps.set_index('uuid')['profile_name'].to_dict()
uuid_to_name_mapper = all_apps.set_index('uuid')['name'].to_dict()
uuid_to_id_mapper = all_apps.set_index('uuid')['id'].to_dict()
all_apps.tail(1)

#all_apps[all_apps['name'].str.contains('Kismet')].T
len(all_apps)

app_dict = {}
for _,app in apps.iterrows():
    recipient = app['recipient'].lower()
    uuid = app['uuid']
    minting_address = app['address'].lower()
    if recipient not in app_dict:
        app_dict.update({
            recipient: {
                'mintingAddressList': [minting_address],
                'uuidList': [uuid]
            }
        })
    else:
        app_dict[recipient]['mintingAddressList'].append(minting_address)
        app_dict[recipient]['uuidList'].append(uuid)

app_json = []        
for recipient, record in app_dict.items():
    app_json.append({
        'recipientAddress': recipient,
        'mintingAddressList': list(set(record['mintingAddressList'])),
        'uuidList': list(set(record['uuidList'])),
    })
        
with open("data/apps/creator_address_to_uuid.json", "w") as f:
    json.dump(app_json, f, indent=2)

farcaster_df = pd.read_parquet('data/raw_metric_data/farcaster.parquet')
fids = farcaster_df.set_index('address')['fid'].to_dict()

## Part 2. Look for NFT deployments from factories

factories_path = 'data/raw_metric_data/oso_raw_creator_factories.csv'
oso_creator_addresses_str = "'" +  "','".join(address_list) + "'"

query = f"""
    select distinct
        originating_address as creator_address,
        contract_address as nft_contract_address,
        lower(network) as blockchain
    from `{PROJECT}.oso.int_factories`
    where originating_address in ({oso_creator_addresses_str})    
"""

result = client.query(query)
print("Query completed.")
factories_df = result.to_dataframe()

print("Dataframe loaded.")
factories_df.set_index('nft_contract_address', drop=True).to_csv(factories_path)

# with open(factories_path) as file:
#     data = file.read()

# table = dune.upload_csv(
#     data=str(data),
#     description="Creator-deployed NFT contracts",
#     table_name="sunny_creator_contracts",
#     is_private=False
# )

factories_df = pd.read_csv(factories_path, index_col=0)
factories_df.tail(1)

## Part 3. Grab NFT mint data from Dune

creator_addresses_str = ',\n\t\t'.join(address_list)
query_sql = f"""

with nft_creators as (
    select
        nft_contract_address,
        blockchain,
        min_by(tx_from, block_number) as creator_address
    from nft.mints
    where
        block_date between date('2024-01-01') and date('2024-09-01')
        and blockchain in ('base', 'optimism', 'zora')
    group by 1,2
    
    union all
    
    select
        nft_contract_address,
        blockchain,
        creator_address
    from dune.gitcoin_dao.dataset_sunny_creator_contracts
)

select
    nft_creators.creator_address,
    nft_creators.nft_contract_address,
    nft_creators.blockchain,
    count(distinct case when block_date > date('2024-06-01') then m.tx_hash end) as num_txns_90D,
    count(distinct m.tx_hash) as num_txns_180D,
    sum(case when block_date > date('2024-06-01') then coalesce(m.amount_usd,0) else 0 end) as amount_usd_90D,
    sum(coalesce(m.amount_usd,0)) as amount_usd_180D,
    array_agg(distinct case when block_date > date('2024-06-01') then m.buyer end) as buyers_90D,
    array_agg(distinct m.buyer) as buyers_180D
from nft.mints as m
join nft_creators
    on m.nft_contract_address = nft_creators.nft_contract_address
    and m.blockchain = nft_creators.blockchain
where
    m.block_date between date('2024-02-01') and date('2024-09-01')
    and nft_creators.creator_address in (
                {creator_addresses_str}
    )
group by 1,2,3

"""

query_id = dune.create_query(name='sunny_nft_mints', query_sql=query_sql, is_private=False)
print(query_id.base.query_id)

query = QueryBase(name='sunny_nft_mints', query_id=query_id.base.query_id)
results_df = dune.run_query_dataframe(query)
results_df.to_parquet('data/raw_metric_data/dune_raw_nft_mints.parquet')

nft_creators = pd.read_parquet('data/raw_metric_data/dune_raw_nft_mints.parquet')

nft_creators['blockchain'] = nft_creators['blockchain'].apply(lambda x: x.title())

def clean_lst(lst_str):
    lst_str = lst_str.replace("[","").replace("]","")
    lst = lst_str.split(" ")
    lst = [x for x in lst if x[:2] == '0x']
    return lst

nft_creators['buyers_90D'] = nft_creators['buyers_90D'].apply(clean_lst)
nft_creators['buyers_180D'] = nft_creators['buyers_180D'].apply(clean_lst)

def map_fids(lst_addr):
    return [x for x in set([fids.get(x) for x in lst_addr]) if x is not None]

nft_creators['list_of_fids_90D'] = nft_creators['buyers_90D'].apply(map_fids)
nft_creators['list_of_fids_180D'] = nft_creators['buyers_180D'].apply(map_fids)

nft_creators.tail(1).T

## Part 4. Derive the metrics

metrics_by_address = pd.concat([
    nft_creators.groupby('creator_address')['nft_contract_address'].nunique().rename('num_drops'),

    nft_creators.groupby('creator_address')['buyers_90D'].agg(lambda x: len(set().union(*x))).rename('num_unique_minters_90D'),
    nft_creators.groupby('creator_address')['buyers_180D'].agg(lambda x: len(set().union(*x))).rename('num_unique_minters_180D'),

    nft_creators.groupby('creator_address')['list_of_fids_90D'].agg(lambda x: len(set().union(*x))).rename('num_farcaster_minters_90D'),
    nft_creators.groupby('creator_address')['list_of_fids_180D'].agg(lambda x: len(set().union(*x))).rename('num_farcaster_minters_180D'),

    nft_creators.groupby('creator_address')['num_txns_90D'].sum().rename('num_transactions_90D'),
    nft_creators.groupby('creator_address')['num_txns_180D'].sum().rename('num_transactions_180D'),
        
    nft_creators.groupby('creator_address')['amount_usd_90D'].sum().rename('usd_value_of_transactions_90D'),
    nft_creators.groupby('creator_address')['amount_usd_180D'].sum().rename('usd_value_of_transactions_180D')
], axis=1)
metrics_by_address

metrics = []
errors = []
for uuid, app in apps.set_index('uuid').iterrows():
    
    name = uuid_to_name_mapper.get(uuid)
    address = app['address']
    recipient = app['recipient']
    
    if address not in metrics_by_address.index:
        errors.append({
            'uuid': uuid,
            **app,
            'status': 'no_nft_activity_90D'
        })
        continue
    
    creator_metrics = metrics_by_address.loc[address].to_dict()
    
    metrics.append({
        'uuid': uuid,
        **app,
        'status': 'metrics_available',
        **creator_metrics,
    })

df_metrics = pd.DataFrame(metrics)
df_metrics.to_csv('data/clean_metric_data/metrics_creators.csv')

## Part 5. Derive metrics by recipient address

recipient_mapping = (
    all_apps
    .groupby(['recipient', 'profile_name'])
    ['address']
    .unique()
    .apply(list)
    .to_dict()
)

recipient_metrics = []
for (recipient,profile_name),address_list in recipient_mapping.items():

    dff = nft_creators[nft_creators.creator_address.isin(address_list)]
    if len(dff) < 1:
        continue

    uuid_list =  all_apps[(all_apps['recipient'] == recipient) & (all_apps['profile_name'] == profile_name)]['uuid'].unique()
    app_id_list = list(set([uuid_to_id_mapper[x] for x in uuid_list]))
    name_list = list(set([uuid_to_name_mapper[x] for x in uuid_list]))
    
    
    nft_creators.groupby('creator_address')['buyers_90D'].agg(lambda x: len(set().union(*x))).rename('num_unique_minters_90D'),
    nft_creators.groupby('creator_address')['buyers_180D'].agg(lambda x: len(set().union(*x))).rename('num_unique_minters_180D'),

    nft_creators.groupby('creator_address')['list_of_fids_90D'].agg(lambda x: len(set().union(*x))).rename('num_farcaster_minters_90D'),
    nft_creators.groupby('creator_address')['list_of_fids_180D'].agg(lambda x: len(set().union(*x))).rename('num_farcaster_minters_180D'),

    nft_creators.groupby('creator_address')['num_txns_90D'].sum().rename('num_transactions_90D'),
    nft_creators.groupby('creator_address')['num_txns_180D'].sum().rename('num_transactions_180D'),
        
    nft_creators.groupby('creator_address')['amount_usd_90D'].sum().rename('usd_value_of_transactions_90D'),
    nft_creators.groupby('creator_address')['amount_usd_180D'].sum().rename('usd_value_of_transactions_180D')

    
    creator_metrics = {
        'recipient': recipient,
        'artist_profile': f'"{profile_name}"',
    
        'num_drops': dff['nft_contract_address'].nunique(),
        
        #'num_unique_minters_90D': len(set(dff['buyers_90D'].sum())),
        'num_unique_minters': len(set(dff['buyers_180D'].sum())),
        
        #'num_farcaster_minters_90D': len(set(dff['list_of_fids_90D'].sum())),
        'num_farcaster_minters': len(set(dff['list_of_fids_180D'].sum())),
        
        #'num_transactions_90D': dff['num_txns_90D'].sum(),
        'num_transactions': dff['num_txns_180D'].sum(),
        
        #'usd_value_of_transactions_90D': dff['amount_usd_90D'].sum(),
        'usd_value_of_transactions': dff['amount_usd_180D'].sum(),
        
        'uuid_list': uuid_list,
        'application_id_list': app_id_list
        #'project_names': name_list
    }
    recipient_metrics.append(creator_metrics)

df_recipient_metrics = pd.DataFrame(recipient_metrics).set_index(['recipient', 'artist_profile'])
df_recipient_metrics.to_csv('data/clean_metric_data/metrics_creators_by_recipient.csv')
df_recipient_metrics.sort_values(by='num_transactions')

## Upload application data to Dune

apps_path = pd.read_csv('data/apps/applications_reviewed.csv')

# with open(apps_path) as file:
#     data = file.read()

# table = dune.upload_csv(
#     data=str(data),
#     description="Sunny Awards applications (cleaned) with reviewer issues.",
#     table_name="sunny_applications",
#     is_private=False
# )
